Transcendence/ft_games/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
cffi
channels==4.1.0
constantly==23.10.4
cryptography
daphne==4.1.2
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
PyYAML==6.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==6.4.post2
django-cors-headers==4.4.0
